Create 8-puzzle.cpp
This C++ file can be run from the command line and will receive as an input an 8-puzzle board (input format is specified at runtime). It will then give the user the options of running a depth or breadth-first search, as well as the famous A* algorithm. This way, they can be compared and one can explore the advantages/disadvantages of each one.